requirements.txt
# Parsonic - Web Scraper GUI
# Core
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
qasync>=0.27.0

# Scraping
httpx>=0.27.0
playwright>=1.40.0
scrapling>=0.2.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Stealth
playwright-stealth>=1.0.0

# Data
pydantic>=2.5.0
pandas>=2.1.0

# Scheduling
APScheduler>=3.10.0

# Database
sqlalchemy>=2.0.0

# Utils
cryptography>=41.0.0
python-dateutil>=2.8.0

# System monitoring
psutil>=5.9.0

# LLM integration (Ollama client)
requests>=2.31.0
